fixed delete. updated xml docs. updated release notes
